fix: FeignClient contextId 명시 (#385)
* fix: FeignClient contextId 명시

* refactor: openfegin client 위치 global로 이동
