+++ support-service/src/main/java/com/node5/supportservice/global/openfeign/client/CatalogClient.java
@@ -1,0 +1,32 @@
+package com.node5.supportservice.global.openfeign.client;
+
+import com.node5.supportservice.global.openfeign.client.dto.ProductIdsRequest;
+import com.node5.supportservice.global.openfeign.client.dto.ProductSummaryListResponse;
+import com.node5.supportservice.global.openfeign.client.dto.ProductStatusResponse;
+import jakarta.validation.Valid;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+
+import java.util.List;
+import java.util.UUID;
+
+@FeignClient(name = "catalog-service", contextId = "catalogClient")
+public interface CatalogClient {
+
+    @GetMapping("/internal/products/getProductsByIds")
+    ResponseEntity<ProductSummaryListResponse> getProductsByIds(
+            @RequestHeader("Member-Id") UUID memberId,
+            @Valid @RequestBody ProductIdsRequest request
+    );
+
+    @GetMapping("/internal/products/ids")
+    ResponseEntity<List<UUID>> getProductIds(Pageable pageable);
+
+    @GetMapping("${api.v1}/products/{productId}/review-status")
+    ProductStatusResponse canPostReview(@PathVariable UUID productId);
+}
