+++ support-service/src/main/java/com/node5/supportservice/global/openfeign/client/dto/ProductStatusResponse.java
@@ -1,4 +1,4 @@
-package com.node5.supportservice.review.client.dto;
+package com.node5.supportservice.global.openfeign.client.dto;
 
 public record ProductStatusResponse(boolean isReviewable) {
 }
