+++ support-service/src/main/java/com/node5/supportservice/global/openfeign/client/dto/ReviewInfo.java
@@ -1,0 +1,4 @@
+package com.node5.supportservice.global.openfeign.client.dto;
+
+public record ReviewInfo() {
+}
