+++ support-service/src/main/java/com/node5/supportservice/recommendation/application/RecommendationService.java
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.node5.supportservice.chat.ChatService;
 import com.node5.supportservice.recommendation.application.dto.ProductRecommendationInfo;
 import com.node5.supportservice.recommendation.application.dto.PromptPayload;
-import com.node5.supportservice.recommendation.client.dto.ProductIdsRequest;
-import com.node5.supportservice.recommendation.client.dto.ProductSummaryListResponse;
-import com.node5.supportservice.recommendation.client.openfeign.OrderClient;
-import com.node5.supportservice.recommendation.client.openfeign.CatalogClient;
+import com.node5.supportservice.global.openfeign.client.dto.ProductIdsRequest;
+import com.node5.supportservice.global.openfeign.client.dto.ProductSummaryListResponse;
+import com.node5.supportservice.global.openfeign.client.OrderClient;
+import com.node5.supportservice.global.openfeign.client.CatalogClient;
 import com.node5.supportservice.recommendation.domain.ProductEmbeddingRepository;
 import com.node5.supportservice.recommendation.exception.RecommendationErrorCode;
 import com.node5.supportservice.recommendation.exception.RecommendationException;
