--- support-service/src/main/java/com/node5/supportservice/review/client/dto/ReviewInfo.java
@@ -1,4 +1,0 @@
-package com.node5.supportservice.review.client.dto;
-
-public record ReviewInfo() {
-}
