+++ support-service/src/main/java/com/node5/supportservice/reviewsummary/batch/ProductIdItemReader.java
@@ -1,6 +1,6 @@
 package com.node5.supportservice.reviewsummary.batch;
 
-import com.node5.supportservice.reviewsummary.client.CatalogClient;
+import com.node5.supportservice.global.openfeign.client.CatalogClient;
 import lombok.RequiredArgsConstructor;
 import org.jspecify.annotations.Nullable;
 import org.springframework.batch.item.ExecutionContext;
